[GraphBolt] Move `get_property` utils.
