+++ python/dgl/graphbolt/impl/sampled_subgraph_impl.py
@@ -6,7 +6,7 @@ from typing import Dict, Union
 import torch
 
 from ..base import CSCFormatBase, etype_str_to_tuple
-from ..internal import get_attributes
+from ..internal_utils import get_attributes
 from ..sampled_subgraph import SampledSubgraph
 
 __all__ = ["SampledSubgraphImpl"]
